requirements.txt
pytorch::pytorch
torchvision
torchaudio
pandas
matplotlib
numpy
g4f